refactor(frontend): split averagingAnalyser.ts into 5 sibling modules
The 698-line averagingAnalyser.ts was well over the 500-line max-lines
guideline (enforced by root eslint.config.js for src/frontend/**).
Split into 5 cohesive files in the same analysers/ directory:

- averagingAnalyser.ts (orchestrator class, 106 lines)
- averagingAnalyser.types.ts (types + factories + accumToMetric, 61 lines)
- averagingAnalyser.filters.ts (assignment filtering helpers, 107 lines)
- averagingAnalyser.accumulation.ts (accumulation + processing, 382 lines)
- averagingAnalyser.rows.ts (perStudent/perTask row builders, 64 lines)

Public API and behaviour of AveragingAnalyser are unchanged. All 27
spec tests pass without modification. computeOverall now takes
criterionWeightings as a parameter (was a class instance method) so
it could be lifted to a module-level function. CriterionWeightings
is now exported from the class file so accumulation.ts can import
the type. No barrel re-exports; helpers are only reachable from
within the analysers/ directory.
